With Api

+++ database/migrations/2024_08_17_194815_create_students_table.php
@@ -6,9 +6,6 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('students', function (Blueprint $table) {
@@ -16,15 +13,11 @@ return new class extends Migration
             $table->string('photo')->nullable();
             $table->string('name');
             $table->string('email')->unique();
-            $table->unsignedInteger('track_id');
+            $table->unsignedBigInteger('track_id');
             $table->foreign('track_id')->references('id')->on('tracks')->onDelete('cascade');
             $table->timestamps();
         });
     }
-
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('students');
